Revert "Common user response interface"
This reverts commit 542de61e29c8354b2d5850fdc9a925b7375666e4.
